% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/replace.R
\name{rr4r_extract}
\alias{rr4r_extract}
\alias{rr4r_extract_all}
\alias{rr4r_extract_groups}
\alias{rr4r_replace}
\title{Extract matching patterns from a string.}
\usage{
rr4r_extract(x, pattern)

rr4r_extract_all(x, pattern)

rr4r_extract_groups(x, pattern)

rr4r_replace(x, pattern, replacement)
}
\arguments{
\item{x}{A character vector to look for the patterns in.}

\item{pattern}{A pattern to look for.}

\item{replacement}{A pattern to replace, or a function to generate the replacement strings.}
}
\description{
Extract matching patterns from a string.

Replace matched patterns in a string.
}
\examples{
### These examples are taken from stringr's doc (?str_extract) ###

shopping_list <- c("apples x4", "bag of flour", "bag of sugar", "milk x2")
rr4r_extract(shopping_list, "\\\\d")
rr4r_extract(shopping_list, "[a-z]+")
rr4r_extract(shopping_list, "[a-z]{1,4}")
rr4r_extract(shopping_list, "\\\\b[a-z]{1,4}\\\\b")

# Extract all matches
rr4r_extract_all(shopping_list, "[a-z]+")
rr4r_extract_all(shopping_list, "\\\\b[a-z]+\\\\b")
rr4r_extract_all(shopping_list, "\\\\d")
}
