% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr4r-package.R
\docType{package}
\name{rr4r-package}
\alias{rr4r}
\alias{rr4r-package}
\title{rr4r: Rust Regex for R}
\description{
Do some regex-fu.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
