% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{rr4r_extract}
\alias{rr4r_extract}
\title{Extract matching patterns from a string.}
\usage{
rr4r_extract(x, pattern)
}
\arguments{
\item{x}{A character vector to look for the patterns in.}

\item{pattern}{A pattern to look for.}
}
\description{
Extract matching patterns from a string.
}
\examples{
shopping_list <- c("apples x4", "bag of flour", "bag of sugar", "milk x2")
rr4r_extract(shopping_list, "\\\\d")
rr4r_extract(shopping_list, "[a-z]+")
rr4r_extract(shopping_list, "[a-z]{1,4}")
rr4r_extract(shopping_list, "\\\\b[a-z]{1,4}\\\\b")
}
